fix mkvmerge

+++ server/internal/adapter/matchtvshow/model.go
@@ -2,7 +2,7 @@ package matchtvshow
 
 type EpisodeInfo struct {
 	// Номер сезона
-	SeasonNumber int
+	SeasonNumber uint8
 	// Номер эпизода в сезоне
 	EpisodeNumber int
 	// Название эпизода, пока не нужно, но может будет нужно для метча в будущем
